chore: updated release config
