+++ .github/workflows/publish-release.yml
@@ -28,20 +28,32 @@ jobs:
         run: pnpm build
 
       - name: Setup NPM for publishing
-        run: cp .npmrc.publish .npmrc
-
-      - name: Publish packages to NPM
-        run: pnpm exec changeset publish
+        run: |
+          echo "//registry.npmjs.org/:_authToken=${NODE_AUTH_TOKEN}" >> .npmrc
+          echo "provenance=true" >> .npmrc
         env:
           NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
+
+      - name: Changesets version and publish
+        uses: changesets/action@v1
+        id: changesets
+        with:
+          version: 'node scripts/version.js'
+          publish: 'pnpm release'
+          createGithubReleases: false
+          title: 'Preparati i pacchetti per il rilascio'
+          commit: 'chore: prepared packages for release'
+        env:
+          NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
           NPM_CONFIG_PROVENANCE: 'true'
 
-      - run: pnpm add -w -D extract-changelog-release
-
       - name: Generate Release Body
+        if: steps.changesets.outputs.published == 'true'
         run: pnpm dlx extract-changelog-release > RELEASE_BODY.md
 
-      - uses: ncipollo/release-action@v1
+      - name: Create GitHub Release
+        if: steps.changesets.outputs.published == 'true'
+        uses: ncipollo/release-action@v1
         with:
           bodyFile: 'RELEASE_BODY.md'
           token: ${{ secrets.GITHUB_TOKEN }}
