+++ package.json
@@ -21,13 +21,8 @@
     "format": "turbo run format",
     "clean": "turbo run clean && rimraf .turbo storybook-static coverage node_modules",
     "prepare": "husky && pnpm test:setup",
-    "release": "turbo run build",
-    "release:changeset": "changeset add",
-    "release:version": "changeset version && pnpm install --lockfile-only",
-    "release:publish": "pnpm build && changeset publish",
-    "release:create": "node scripts/release.js",
-    "release:changelog": "node scripts/generate-unified-changelog.js",
-    "release:version-bump": "node scripts/add-version-bump.js",
+    "release": "pnpm build && changeset publish",
+    "changeset": "changeset",
     "test:setup": "pnpm exec playwright install",
     "test:setup:ci": "pnpm exec playwright install --with-deps --only-shell",
     "test": "turbo run test",
