+++ packages/typescript-config/package.json
@@ -1,4 +1,5 @@
 {
   "name": "@italia/typescript-config",
-  "version": "0.0.0"
+  "version": "0.0.0",
+  "private": true
 }
